+++ examples/pidof.py
@@ -7,7 +7,7 @@
 
 """
 A clone of 'pidof' cmdline utility.
-$ pidof  /usr/bin/python
+$ pidof python
 1140 1138 1136 1134 1133 1129 1127 1125 1121 1120 1119
 """
 
@@ -16,28 +16,31 @@ import psutil
 import sys
 
 
-def pidsof(pgm):
+def pidof(pgname):
     pids = []
-    # Iterate on all proccesses and find matching cmdline and get list of
-    # corresponding PIDs
     for proc in psutil.process_iter():
+        # search for matches in the process name and cmdline
+        try:
+            name = proc.name()
+        except psutil.Error:
+            name = ""
         try:
             cmdline = proc.cmdline()
-            pid = proc.pid
         except psutil.Error:
-            continue
-        if len(cmdline) > 0 and cmdline[0] == pgm:
-            pids.append(str(pid))
+            cmdline = []
+        if name == pgname or cmdline and cmdline[0] == pgname:
+            pids.append(str(proc.pid))
     return pids
 
 
 def main():
     if len(sys.argv) != 2:
-        sys.exit('usage: %s pgm_name' % __file__)
+        sys.exit('usage: %s pgname' % __file__)
     else:
-        pgmname = sys.argv[1]
-    pids = pidsof(pgmname)
-    print(" ".join(pids))
+        pgname = sys.argv[1]
+    pids = pidof(pgname)
+    if pids:
+        print(" ".join(pids))
 
 if __name__ == '__main__':
     main()